requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn
jinja2
python-multipart
pyttsx3
itsdangerous